Bag of hacks that seems to work

+++ joblib/_parallel_backends.py
@@ -644,6 +644,9 @@ class LokyBackend(AutoBatchingMixin, ParallelBackendBase):
     def abort_everything(self, ensure_ready=True):
         """Shutdown the workers and restart a new one with the same parameters
         """
+        if self._workers is None:
+            return
+
         self._workers.terminate(kill_workers=True)
         self._workers = None
